GUI/requirements.txt
numpy
scipy
pandas
opencv-python
tensorflow-macos
tensorflow-metal
matplotlib
shapely
scikit-image
PyWavelets
ttkbootstrap